profile-service/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
boto3==1.35.42
botocore==1.35.42
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
Django==5.0.8
django-cors-headers==4.4.0
django-environ==0.11.2
django-keycloak==0.1.1
django-keycloak-auth==1.0.0
django-minio-storage==0.5.7
django-oauth-toolkit==2.4.0
django-recaptcha==4.0.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
ecdsa==0.19.0
environ==1.0
et-xmlfile==1.1.0
executing==2.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
ipython==8.27.0
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
matplotlib-inline==0.1.7
minio==7.2.9
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak-client==0.2.3
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.7.0
